+++ tests/UnitTests/Curiosity.Migrations.UnitTests/DbVersion_Should.cs
@@ -11,7 +11,7 @@ public class DbVersion_Should
     [Fact]
     public void TryParse_Numeric_Major_Version()
     {
-        var parseResult = DbVersion.TryParse("1", out var version);
+        var parseResult = MigrationVersion.TryParse("1", out var version);
 
         parseResult.Should().BeTrue("it's correct version");
         version.Major.Should().Be(1, "version contains only major version");
@@ -21,7 +21,7 @@ public class DbVersion_Should
     [Fact]
     public void Pattern_Should_Handle_Numeric_Major_Minor_Version()
     {
-        var parseResult = DbVersion.TryParse("1.2", out var version);
+        var parseResult = MigrationVersion.TryParse("1.2", out var version);
 
         parseResult.Should().BeTrue("it's correct version");
         version.Major.Should().Be(1, "version contains major version");
@@ -31,7 +31,7 @@ public class DbVersion_Should
     [Fact]
     public void Pattern_Should_Handle_Numeric_Major_Minor_2_Version()
     {
-        var parseResult = DbVersion.TryParse("10001.020", out var version);
+        var parseResult = MigrationVersion.TryParse("10001.020", out var version);
 
         parseResult.Should().BeTrue("it's correct version");
         version.Major.Should().Be(10001, "version contains major version");
@@ -41,7 +41,7 @@ public class DbVersion_Should
     [Fact]
     public void Pattern_Should_Handle_Yyymmdd_Version()
     {
-        var parseResult = DbVersion.TryParse("20201208", out var version);
+        var parseResult = MigrationVersion.TryParse("20201208", out var version);
 
         parseResult.Should().BeTrue("it's correct version");
         version.Major.Should().Be(20201208, "version contains major version");
@@ -51,7 +51,7 @@ public class DbVersion_Should
     [Fact]
     public void Pattern_Should_Handle_Yyymmdd_hhmm_Version()
     {
-        var parseResult = DbVersion.TryParse("20201208-1850", out var version);
+        var parseResult = MigrationVersion.TryParse("20201208-1850", out var version);
 
         parseResult.Should().BeTrue("it's correct version");
         version.Major.Should().Be(202012081850, "version contains major version");
@@ -61,7 +61,7 @@ public class DbVersion_Should
     [Fact]
     public void Pattern_Should_Handle_Yyymmdd_hhmm_minor_Version()
     {
-        var parseResult = DbVersion.TryParse("20201208-1850.02", out var version);
+        var parseResult = MigrationVersion.TryParse("20201208-1850.02", out var version);
 
         parseResult.Should().BeTrue("it's correct version");
         version.Major.Should().Be(202012081850, "version contains major version");